requirements.txt
python-binance
PySimpleGUI